feat(extensions): typed context keys for non-product hook inputs
Promotes ExtensionContextKey<T> + ExtensionContextData from
:data-render-compose to :data-render-core so post-capture extensions can
declare typed handoffs for render-host-controlled inputs (Android View
root, future scope, etc.) instead of using the string-keyed
attributes: Map<String, Any?> escape hatch.

ExtensionPostCaptureContext gains a `data: ExtensionContextData` field
alongside the legacy `attributes` map. Convenience get/require shorthands
mirror the Compose-side ExtensionComposeContext.

AccessibilityHierarchyExtension switches over: declares
AccessibilityHierarchyContextKeys.ViewRoot: ExtensionContextKey<View>
and reads it via context.require(...). Drops the
VIEW_ROOT_ATTRIBUTE string constant. The Android-platform-binding
(View as the payload type) lives in :data-a11y-hierarchy-android — a
future :data-a11y-hierarchy-desktop would declare its own ViewRoot key
with the Desktop equivalent type, and downstream consumers stay
type-safe within the platform variant they're paired with.

Cross-module imports (:data-recomposition-connector, the render-compose
test) updated to import the moved types from the render-core package.
The Compose hook surface (ExtensionComposeContext et al.) imports the
same types unchanged — back-compat preserved on the Compose side.

Test plan
- :data-render-core, :data-render-compose, :data-recomposition-connector,
  :data-a11y-hierarchy-android, :data-a11y-connector, :daemon:android,
  :renderer-android, :data-theme-connector all compile and tests pass.
- failsLoudlyWhenViewRootContextKeyMissing now asserts on the typed
  key's name instead of the deprecated attribute constant.
