+++ data/render/core/src/main/kotlin/ee/schimke/composeai/data/render/extensions/PostCaptureProcessor.kt
@@ -5,14 +5,24 @@ package ee.schimke.composeai.data.render.extensions
  * captured. No Compose runtime, no platform-specific types — extensions read declared inputs and
  * write declared outputs through [products], which the executor scopes to the extension's
  * `inputs`/`outputs` set so contract violations fail loudly at runtime.
+ *
+ * Non-product context inputs the host hands to extensions go through [data] (typed
+ * [ExtensionContextKey] entries) for new extensions. The legacy [attributes] map is kept for
+ * back-compat with extensions that haven't migrated to typed keys yet (overlay attribute slots in
+ * particular); prefer [data] in new code.
  */
 data class ExtensionPostCaptureContext(
   val extensionId: DataExtensionId,
   val previewId: String?,
   val renderMode: String?,
   val products: DataProductStore,
+  val data: ExtensionContextData = ExtensionContextData.Empty,
   val attributes: Map<String, Any?> = emptyMap(),
-)
+) {
+  fun <T : Any> get(key: ExtensionContextKey<T>): T? = data.get(key)
+
+  fun <T : Any> require(key: ExtensionContextKey<T>): T = data.require(key)
+}
 
 /**
  * Hook for extensions that run once after a render's bitmap is captured. Distinct from
